+++ crates/insights/src/cli/mod.rs
@@ -1,3 +1,4 @@
 pub mod client;
 pub mod commands;
+pub mod display;
 pub mod server_manager;
