+++ crates/insights/src/lib.rs
@@ -4,9 +4,4 @@
 //! storage and retrieval for development workflows and team collaboration.
 
 pub mod cli;
-pub mod gte_base;
-pub mod insight;
-pub mod search;
-pub mod semantic;
 pub mod server;
-pub mod similarity;
