--- crates/insights/src/semantic.rs
@@ -1,3 +1,0 @@
-//! Semantic analysis utilities
-//!
-//! Note: Semantic similarity functionality is implemented in the similarity module.
