+++ crates/insights/src/server/mod.rs
@@ -5,6 +5,8 @@
 
 pub mod handlers;
 pub mod middleware;
+pub mod models;
 pub mod routing;
 pub mod server;
+pub mod services;
 pub mod types;
