+++ crates/insights/src/server/models/insight.rs
