+++ crates/insights/src/server/services/embeddings.rs
