+++ crates/insights/src/server/services/mod.rs
@@ -1,0 +1,3 @@
+pub mod embeddings;
+pub mod search;
+pub mod similarity;
