+++ crates/insights/src/server/services/similarity.rs
